Proximity kernel module, to tweak and read real sensor data.
Allow to retrieve real proximity registers, not only the on-off
used to switch on/off the lcd screen when we make a call.

Beware : IR LED uses a lot of power (110mA or 220mA) by pulses or not
         The constant mode is "locked" because that could alter the LED

This kernel module is not enabled for the moment, but can help later
to enhance our device. !! NOT COMPATIBLE WITH DEFY PLUS !!